Moving rate files
