+++ src/app/rate/flat-rate.spec.ts
