+++ src/app/rate/rate.ts
